Move Factorials in integer.combo
